Fix the delay in the test
