+++ java-util/src/main/java/io/druid/java/util/common/granularity/Granularities.java
@@ -2,19 +2,20 @@ package io.druid.java.util.common.granularity;
 
 public class Granularities
 {
-  public static final Granularity SECOND = Granularity.GranularityType.SECOND.getDefaultGranularity();
-  public static final Granularity MINUTE = Granularity.GranularityType.MINUTE.getDefaultGranularity();
-  public static final Granularity FIVE_MINUTE = Granularity.GranularityType.FIVE_MINUTE.getDefaultGranularity();
-  public static final Granularity TEN_MINUTE = Granularity.GranularityType.TEN_MINUTE.getDefaultGranularity();
-  public static final Granularity FIFTEEN_MINUTE = Granularity.GranularityType.FIFTEEN_MINUTE.getDefaultGranularity();
-  public static final Granularity THIRTY_MINUTE = Granularity.GranularityType.THIRTY_MINUTE.getDefaultGranularity();
-  public static final Granularity HOUR = Granularity.GranularityType.HOUR.getDefaultGranularity();
-  public static final Granularity SIX_HOUR = Granularity.GranularityType.SIX_HOUR.getDefaultGranularity();
-  public static final Granularity DAY = Granularity.GranularityType.DAY.getDefaultGranularity();
-  public static final Granularity WEEK = Granularity.GranularityType.WEEK.getDefaultGranularity();
-  public static final Granularity MONTH = Granularity.GranularityType.MONTH.getDefaultGranularity();
-  public static final Granularity QUARTER = Granularity.GranularityType.QUARTER.getDefaultGranularity();
-  public static final Granularity YEAR = Granularity.GranularityType.YEAR.getDefaultGranularity();
-  public static final Granularity ALL = Granularity.GranularityType.ALL.getDefaultGranularity();
-  public static final Granularity NONE = Granularity.GranularityType.NONE.getDefaultGranularity();
+  public static final Granularity SECOND = GranularityType.SECOND.getDefaultGranularity();
+  public static final Granularity MINUTE = GranularityType.MINUTE.getDefaultGranularity();
+  public static final Granularity FIVE_MINUTE = GranularityType.FIVE_MINUTE.getDefaultGranularity();
+  public static final Granularity TEN_MINUTE = GranularityType.TEN_MINUTE.getDefaultGranularity();
+  public static final Granularity FIFTEEN_MINUTE = GranularityType.FIFTEEN_MINUTE.getDefaultGranularity();
+  public static final Granularity THIRTY_MINUTE = GranularityType.THIRTY_MINUTE.getDefaultGranularity();
+  public static final Granularity HOUR = GranularityType.HOUR.getDefaultGranularity();
+  public static final Granularity SIX_HOUR = GranularityType.SIX_HOUR.getDefaultGranularity();
+  public static final Granularity DAY = GranularityType.DAY.getDefaultGranularity();
+  public static final Granularity WEEK = GranularityType.WEEK.getDefaultGranularity();
+  public static final Granularity MONTH = GranularityType.MONTH.getDefaultGranularity();
+  public static final Granularity QUARTER = GranularityType.QUARTER.getDefaultGranularity();
+  public static final Granularity YEAR = GranularityType.YEAR.getDefaultGranularity();
+  public static final Granularity ALL = GranularityType.ALL.getDefaultGranularity();
+  public static final Granularity NONE = GranularityType.NONE.getDefaultGranularity();
+
 }
