+++ java-util/src/test/java/io/druid/java/util/common/GranularityTest.java
@@ -21,6 +21,7 @@ package io.druid.java.util.common;
 
 import io.druid.java.util.common.granularity.Granularities;
 import io.druid.java.util.common.granularity.Granularity;
+import io.druid.java.util.common.granularity.GranularityType;
 import io.druid.java.util.common.granularity.PeriodGranularity;
 import org.joda.time.DateTime;
 import org.joda.time.DateTimeZone;
@@ -413,7 +414,7 @@ public class GranularityTest {
     {
         try {
           Period p = Period.years(6).withMonths(3).withSeconds(23);
-          Granularity.GranularityType.fromPeriod(p);
+          GranularityType.fromPeriod(p);
           Assert.fail("Complicated period creation should fail b/c of unsupported granularity type.");
         }
         catch (IAE e) {
